requirements.txt
pyTelegramBotAPI~=4.14.0
python-dotenv~=0.18.0
redis~=5.0.1